refactor(server): split ws/message_service into per-concern submodules
